authentication added

+++ base/views.py
@@ -1,7 +1,62 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+
 from .forms import RoomForm
 from .models import Room, Topic
+
+def userLogin(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password does not match")
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def userLogout(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False) #access the user right away but don't commit yet
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -15,6 +70,7 @@ def room(request,pk):
     room = Room.objects.get()
     return render(request, 'base/room.html')
 
+@login_required(login_url='login')
 def roomForm(request):
     if request.method == 'POST':
         form = RoomForm(request.POST)
@@ -24,9 +80,14 @@ def roomForm(request):
     form = RoomForm()
     return render(request, 'base/room_form.html', {'form': form})
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -34,9 +95,15 @@ def updateRoom(request, pk):
             return redirect('home')
     return render(request, 'base/room_form.html', {'form': form})
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html',{'obj': room})
+
